src/web/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
ultralytics==8.0.196
werkzeug==2.3.7
python-magic==0.4.27
